requirements.txt
flask
discord.py
discord.py[voice]
python-dotenv
requests
asyncio
aiosqlite
youtube_dl
